drivers: qcom: qfprom: move MX rail workaround into per-platform config
QFPROM fuse blowing on Kodiak requires toggling the MX voltage rail
via RPMH/CMD_DB around the operation. Other platforms, such as
Lemans, need no such workaround, and future chipsets may or may not
need it either.

This was previously handled with a CFG_QFPROM_MX_RAIL_WA build flag
selecting between a real implementation and stub functions, which
required #ifdef-driven build plumbing and would not scale cleanly to
additional platforms.

Replace this with optional init()/deinit() callbacks in struct
qfprom_platform_config. Kodiak's platform config populates them with
its RPMH/CMD_DB-based MX rail handling; platforms that need no
platform-specific init/deinit simply leave them unset. qfprom_hw_init()
and qfprom_hw_deinit() invoke them through NULL-guarded calls, so no
CFG flag or #ifdef is needed to add or drop the workaround per
platform.
